Extract session logic to it's own module
