+++ packages/account_sdk/src/lib.rs
@@ -8,10 +8,12 @@ pub mod factory;
 pub mod hash;
 pub mod paymaster;
 pub mod provider;
+pub mod session;
 pub mod signers;
 pub mod storage;
 mod transaction_waiter;
 pub mod typed_data;
+pub mod utils;
 
 #[cfg(not(target_arch = "wasm32"))]
 #[cfg(test)]
